+++ libs/angular/config/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.3](https://github.com/tractr/stack/compare/angular-config-0.0.2...angular-config-0.0.3) (2021-07-15)
+
+
+### Bug Fixes
+
+* angular config build correctly ([#190](https://github.com/tractr/stack/issues/190)) ([bd8265b](https://github.com/tractr/stack/commit/bd8265bbf6f02c76e3f94242dce9b18cf5cd3653))
+* fix bin name export and add bin node inside the index file ([#188](https://github.com/tractr/stack/issues/188)) ([04237e1](https://github.com/tractr/stack/commit/04237e19c611af400adbd0cef37e402304d41509))
+
+
+
 ## [0.0.2](https://github.com/tractr/stack/compare/angular-config-0.0.1...angular-config-0.0.2) (2021-07-15)